+++ mainapp/templates/mainapp/main_view.html
@@ -4,7 +4,7 @@
 
 <head>
     <meta charset="UTF-8">
-    <title>WhatToWatch.tv</title>
+    <title>MovieBreak.tv</title>
 
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
@@ -16,7 +16,7 @@
 
         <header class="main-header">
             <div class="logotype">
-                <h1 class="logotext">MoviePick</h1>
+                <h1 class="logotext">MovieBreak</h1>
                 <!-- <img class="logo" src="{% static 'logo.png' %}" alt="Logo"> -->
                 {{processing_time}}
             </div>
@@ -26,7 +26,7 @@
         {% if welcome_message %}
         <div class="welcome-mess-div">
             <div class="welcome-mess-div-2">
-                <h1 class="white-text">Meet MoviePick, minimalistic movie recommender fueled with AI.</h1>
+                <h1 class="white-text">Meet MovieBreak, minimalistic movie recommender fueled with AI.</h1>
                 <br>
                 <br>
                 <h5 class="white-text"> MovieNeon, the intelligent movie matchmaker, is at your service. Share your
