+++ mainapp/views.py
@@ -1,15 +1,15 @@
 from django.shortcuts import render
+from django import forms
+
 from .forms import UserPrompt
+from .utils import get_movie_info_tmdb
+from .models import Movie
+
 from openai import OpenAI
 import time
 import ast
-import movieposters as mp
-import requests
-from bs4 import BeautifulSoup
-from .models import Movie
 from datetime import datetime
-from .utils import check_platforms_for_a_movie
-from django import forms
+
 
 def main_view(request):
     client = OpenAI()   
@@ -27,18 +27,18 @@ def main_view(request):
 
             if prompt_form.cleaned_data["gpt_4"]:
                 assistant_id = 'asst_6bqZAqHpKP48jTaxrCSlbbxL' # 'asst_sKuRYRpYQWM6VigUHTnFzUhk'
-                file_id = 'file-E7IBRtoF7uImZmH9kU36urA1'
+                # file_id = 'file-E7IBRtoF7uImZmH9kU36urA1'
                 tools = [{"type": "retrieval"}]
                 file_ids = [file_id]
             if prompt_form.cleaned_data["only_watchlist"]:
                 assistant_id = 'asst_sKuRYRpYQWM6VigUHTnFzUhk'
-                file_id = 'file-6J32Lw7YXCBQyIuYCA1tRAsH'
+                # file_id = 'file-6J32Lw7YXCBQyIuYCA1tRAsH'
                 tools = [{"type": "retrieval"}]
                 file_ids = [file_id]
 
             elif prompt_form.cleaned_data["without_seen"]:
                 assistant_id = 'asst_sKuRYRpYQWM6VigUHTnFzUhk'
-                file_id = 'file-qWa8QHBYup9GjeskZCHtvlFh'
+                # file_id = 'file-qWa8QHBYup9GjeskZCHtvlFh'
                 tools = [{"type": "retrieval"}]
                 file_ids = [file_id]
 
@@ -54,7 +54,7 @@ def main_view(request):
                 pass
             else:
                 assistant_id = 'asst_6bqZAqHpKP48jTaxrCSlbbxL'#'asst_sKuRYRpYQWM6VigUHTnFzUhk'
-                file_id = 'file-E7IBRtoF7uImZmH9kU36urA1'
+                # file_id = 'file-E7IBRtoF7uImZmH9kU36urA1'
                 tools = [{"type": "retrieval"}]
                 file_ids = [file_id]
                 prompt_additional_info = f' available on of these streaming platforms: {selected_platforms} .Please remember to not write any additional text in a response, provide just a list.'
@@ -84,32 +84,45 @@ def main_view(request):
             end_index = string_data.rfind(']') + 1
             movies_list_str = string_data[start_index:end_index]
             response = ast.literal_eval(movies_list_str)
-            print(f"Before imdb data {time.time() - start_time}")
+            print(f"Before tmdb data {time.time() - start_time}")
             for movie in response:
                 movie_title = movie["Title"]
-                db_movie = Movie.objects.filter(title=movie_title, year=movie["Year"])
+                movie_year = str(movie["Year"])
+                db_movie = Movie.objects.filter(title=movie_title, year=movie_year)
                 if db_movie.exists():
-                    db_movie = Movie.objects.get(title=movie_title, year=movie["Year"])
+                    db_movie = Movie.objects.get(title=movie_title, year=movie_year)
                     poster_link = db_movie.poster_url
                     length = db_movie.length
-                    movie_link = db_movie.imdb_link
+                    movie_link = db_movie.tmdb_link
                     rating = db_movie.rating
+                    try:
+                        streaming = db_movie.streaming_services["US"]["flatrate"]
+                        print(streaming)
+                        streaming_services = [f"https://image.tmdb.org/t/p/original{platform['logo_path']}" for platform in streaming]
+                        print(streaming_services)
+                    except:
+                        streaming_services = []
                 else:
+                    print(movie)
                     print("NOT EXISTS")
-                    imdb_link_and_local_poster = mp.get_imdb_link_from_title(movie_title)
-                    movie_link = imdb_link_and_local_poster[0]
-                    poster_rating_length = mp.get_poster_from_imdb_link(imdb_link_and_local_poster[0])
-                    poster_link = poster_rating_length[0]
-                    rating = poster_rating_length[1]
-                    length = poster_rating_length[2]
-                    streaming_platforms = check_platforms_for_a_movie(movie_title, movie["Year"])
-                    db_movie = Movie(title=movie_title, year=movie["Year"], length=length, poster_url = poster_link, imdb_link=imdb_link_and_local_poster[0], rating=rating, netflix = streaming_platforms[0], amazon_prime= streaming_platforms[1], hulu= streaming_platforms[2], disney_plus= streaming_platforms[3], hbo_max= streaming_platforms[4], apple_tv= streaming_platforms[5], peacock= streaming_platforms[6], last_update=datetime.today().date())
+                    movie_info = get_movie_info_tmdb(movie_title, movie_year)
+                    movie_link = movie_info["TMDB link"]
+                    poster_link = movie_info["Poster"]
+                    rating = movie_info["Rating"]
+                    length = movie_info["Length"]
+                    try:
+                        print(movie_info["Streaming"]["US"]["flatrate"])
+                        streaming_services = [f"https://image.tmdb.org/t/p/original{platform['logo_path']}" for platform in movie_info["Streaming"]["US"]["flatrate"]]
+                        print(streaming_services)
+                    except:
+                        streaming_services = []
+                    db_movie = Movie(title=movie_title, year=movie["Year"], length=length, poster_url = poster_link, tmdb_link=movie_link, streaming_services = movie_info["Streaming"], rating=rating, last_update=datetime.today().date())
                     db_movie.save()
                 movie["Rating"] = rating
                 movie["Poster"] = poster_link
                 movie["Link"] = movie_link
                 movie["Length"] = length
-                movie["Platforms"] = db_movie.get_streaming_platforms()
+                movie["Platforms"] = streaming_services
                 movie["Description"] = movie["Plot short description"]
             processing_time = time.time() - start_time
             prompt_form = UserPrompt()
